Fix ECAM

+++ devicesd/include/pci/pci.h
@@ -135,10 +135,10 @@ extern PCIDeviceVector pci_devices;
 
 uint32_t pci_read_register(struct PCIDevicePtr *s, unsigned register);
 void pci_write_register(struct PCIDevicePtr *s, unsigned register, uint32_t value);
-int fill_device(struct PCIDevicePtr *s, struct PCIHostBridge *g, int bus, int device, int function);
+int fill_device(struct PCIDevicePtr *s, struct PCIHostBridge *g, uint8_t bus, uint8_t device, uint8_t function);
 
 extern bool pci_fully_working;
-int fill_device_early(struct PCIDevicePtr *s, int bus, int device, int function);
+int fill_device_early(struct PCIDevicePtr *s, uint8_t bus, uint8_t device, uint8_t function);
 
 inline uint16_t pci_read_word(struct PCIDevicePtr *s, unsigned offset)
 {
